// Template.Client/src/models/generated.ts
export enum UserStatus {
    Active = "Active",
    Inactive = "Inactive",
    Pending = "Pending",
    Suspended = "Suspended",
}